Organizes all the LTC6811 related python simulations into ltc6811.py
