+++ BSP/Simulator/DataGeneration/SPI.py
@@ -3,6 +3,7 @@ import random
 import os
 import fcntl
 import config
+import ltc6811
 
 """
 This module generates Voltage and Temperature
@@ -19,60 +20,6 @@ temperature_values = []     # list of 31 pairs of sensors (fixed point 0.001)
 
 # LTC6811 objects
 minions = []
-
-# dictionary of all command codes
-command_codes = {
-    'SIM_LTC6811_WRCFGA'    : {'code': 0x001, 'handler': None},
-    'SIM_LTC6811_WRCFGB'    : {'code': 0x024, 'handler': None},
-    'SIM_LTC6811_RDCFGA'    : {'code': 0x002, 'handler': None},
-    'SIM_LTC6811_RDCFGB'    : {'code': 0x026, 'handler': None},
-    'SIM_LTC6811_RDCVA'     : {'code': 0x004, 'handler': None},
-    'SIM_LTC6811_RDCVB'     : {'code': 0x006, 'handler': None},
-    'SIM_LTC6811_RDCVC'     : {'code': 0x008, 'handler': None},
-    'SIM_LTC6811_RDCVD'     : {'code': 0x00A, 'handler': None},
-    'SIM_LTC6811_RDCVE'     : {'code': 0x009, 'handler': None},
-    'SIM_LTC6811_RDCVF'     : {'code': 0x00B, 'handler': None},
-    'SIM_LTC6811_RDAUXA'    : {'code': 0x00C, 'handler': None},
-    'SIM_LTC6811_RDAUXB'    : {'code': 0x00E, 'handler': None},
-    'SIM_LTC6811_RDAUXC'    : {'code': 0x00D, 'handler': None},
-    'SIM_LTC6811_RDAUXD'    : {'code': 0x00F, 'handler': None},
-    'SIM_LTC6811_RDSTATA'   : {'code': 0x010, 'handler': None},
-    'SIM_LTC6811_RDSTATB'   : {'code': 0x012, 'handler': None},
-    'SIM_LTC6811_WRSCTRL'   : {'code': 0x014, 'handler': None},
-    'SIM_LTC6811_WRPWM'     : {'code': 0x020, 'handler': None},
-    'SIM_LTC6811_WRPSB'     : {'code': 0x01C, 'handler': None},
-    'SIM_LTC6811_RDSCTRL'   : {'code': 0x016, 'handler': None},
-    'SIM_LTC6811_RDPWM'     : {'code': 0x022, 'handler': None},
-    'SIM_LTC6811_RDPSB'     : {'code': 0x01E, 'handler': None},
-    'SIM_LTC6811_STSCTRL'   : {'code': 0x019, 'handler': None},
-    'SIM_LTC6811_CLRSCTRL'  : {'code': 0x018, 'handler': None},
-    'SIM_LTC6811_ADCV'      : {'code': 0x260, 'handler': None},       # 0 1 MD[1] MD[0] 1 1 DCP 0 CH[2] CH[1] CH[0]
-    'SIM_LTC6811_ADOWPU'    : {'code': 0x268, 'handler': None},       # 0 1 MD[1] MD[0] PUP 1 DCP 1 CH[2] CH[1] CH[0]
-    'SIM_LTC6811_ADOWPD'    : {'code': 0x228, 'handler': None},       # 0 1 MD[1] MD[0] PUP 1 DCP 1 CH[2] CH[1] CH[0]
-    'SIM_LTC6811_CVST'      : {'code': 0x207, 'handler': None},       # 0 1 MD[1] MD[0] ST[1] ST[0] 0 0 1 1 1
-    'SIM_LTC6811_ADOL'      : {'code': 0x201, 'handler': None},       # 0 1 MD[1] MD[0] 0 0 DCP 0 0 0 1
-    'SIM_LTC6811_ADAX'      : {'code': 0x460, 'handler': None},       # 1 0 MD[1] MD[0] 1 1 0 0 CHG[2] CHG[1] CHG[0]
-    'SIM_LTC6811_ADAXD'     : {'code': 0x400, 'handler': None},       # 1 0 MD[1] MD[0] 0 0 0 0 CHG[2] CHG[1] CHG[0]
-    'SIM_LTC6811_AXST'      : {'code': 0x407, 'handler': None},       # 1 0 MD[1] MD[0] ST[1] ST[0] 0 0 1 1 1
-    'SIM_LTC6811_ADSTAT'    : {'code': 0x468, 'handler': None},       # 1 0 MD[1] MD[0] 1 1 0 1 CHST[2] CHST[1] CHST[0]
-    'SIM_LTC6811_ADSTATD'   : {'code': 0x468, 'handler': None},       # 1 0 MD[1] MD[0] 0 0 0 1 CHST[2] CHST[1] CHST[0]
-    'SIM_LTC6811_STATST'    : {'code': 0x40F, 'handler': None},       # 1 0 MD[1] MD[0] ST[1] ST[0] 0 1 1 1 1
-    'SIM_LTC6811_ADCVAX'    : {'code': 0x46F, 'handler': None},       # 1 0 MD[1] MD[0] 1 1 DCP 1 1 1 1
-    'SIM_LTC6811_ADCVSC'    : {'code': 0x467, 'handler': None},       # 1 0 MD[1] MD[0] 1 1 DCP 0 1 1 1
-    'SIM_LTC6811_CLRCELL'   : {'code': 0x711, 'handler': None},
-    'SIM_LTC6811_CLRAUX'    : {'code': 0x712, 'handler': None},
-    'SIM_LTC6811_CLRSTAT'   : {'code': 0x713, 'handler': None},
-    'SIM_LTC6811_PLADC'     : {'code': 0x714, 'handler': None},
-    'SIM_LTC6811_DIAGN'     : {'code': 0x715, 'handler': None},
-    'SIM_LTC6811_WRCOMM'    : {'code': 0x721, 'handler': None},
-    'SIM_LTC6811_RDCOMM'    : {'code': 0x722, 'handler': None},
-    'SIM_LTC6811_STCOMM'    : {'code': 0x723, 'handler': None}
-}
-
-mux_addresses = {
-    'SIM_LTC1380_MUX1' : 0x90,        # MUX addresses on Minion boards
-    'SIM_LTC1380_MUX2' : 0x92
-}
 
 def open_wires(battery=None):
     """
@@ -198,115 +145,6 @@ def read():
     return values
 
 
-class LTC6811:
-    CRC15_POLY = 0x4599
-    pec15_table = [0] * 256
-
-
-    def __init__(self, batt_module):
-        self.batt_modules = batt_modules
-
-        # reg suffixes indicate data stored in LTC6811 registers
-        self.config_reg = []    # 6B: Configuration register of LTC
-
-        # staging buffers that will be used when formatting bus
-        self.curr_cmd = 0       # 12-bits
-        self.tx_data = []       # 6B: POV of simulator
-        self.rx_data = []       # 6B
-
-        # As long this is called once, pec will work
-        LTC6811.pec_init()
-
-
-    def __str__(self):
-        print("Config: ", end='')
-        for byte in self.config_reg:
-            print(hex(byte), end=' ')
-
-
-    def command_update(self, command, data):
-        self.curr_cmd = command
-        self.rx_data = data
-
-
-
-    @staticmethod
-    def format_full_protocol(ics):
-        protocol = []
-
-        for ic in ics:
-            # Copy data
-            for byte in ic.tx_data:
-                protocol.append(byte)
-            
-            pec = LTC6811.calc_pec(ic.tx_data)
-            protocol.append((pec >> 8) & 0x00FF)    # PEC0
-            protocol.append(pec & 0x00FF)           # PEC1
-
-        return protocol
-
-
-    @staticmethod
-    def parse_full_protocol(ics, protocol):
-        DATA_START_IDX = 4
-
-        # First to bytes are the command
-        # The next two lines of code pops the first 4 bytes which are the cmd
-        #   related bytes.
-        cmd = ((protocol.pop(0) << 8) & 0xFF00) | (protocol.pop(0) & 0x00FF)
-        cmd_pec = ((protocol.pop(0) << 8) & 0xFF00) | (protocol.pop(0) & 0x00FF)
-
-        # Update all the LTCs with new cmd
-        for ic in ics:
-            data = [0] * 6
-            for i in range(6):
-                data[i] = protocol.pop(0)   # Pop the data
-
-            # data_pec saved but not needed
-            data_pec = ((protocol.pop(0) << 8) & 0xFF00) | (protocol.pop(0) & 0x00FF)
-
-            # Update LTC
-            ic.command_update(cmd, data)
-
-
-    @staticmethod
-    def print_protocol(protocol):
-        print()
-        print("Idx : Data")
-        idx = 0
-        for byte in protocol:
-            print(" {0:2} : {1}".format(idx, hex(protocol[idx])))
-            idx += 1
-
-
-    @staticmethod
-    def pec_init():
-        for i in range(256):
-            remainder = i << 7
-            for bit in range(8, 0, -1):
-                if remainder & 0x4000:
-                    remainder = ((remainder << 1))
-                    remainder = (remainder ^ LTC6811.CRC15_POLY)
-                else:
-                    remainder = ((remainder << 1))
-            LTC6811.pec15_table[i] = (remainder&0x0000FFFF)
-
-
-    @staticmethod
-    def calc_pec(bytes):
-        """
-        @brief  calculates 16-bit PEC value of the byte array
-        @param  bytes : array of bytes that the algorithm will use to calculate
-                        PEC. For the LTC6811, the length of the array is either
-                        2 (command) or 6 (data) bytes long
-        """
-        address = 0
-        remainder = 16  # PEC seed
-        for i in range(len(bytes)):
-            address = ((remainder >> 7) ^ bytes[i]) & 0xff   #calculate PEC table address
-            remainder = (remainder << 8 ) ^ LTC6811.pec15_table[address]
-        return (remainder*2 & 0x0000FFFF) # The CRC15 has a 0 in the LSB so the final value must be multiplied by 2
-
 
 
 
@@ -337,28 +175,32 @@ if __name__ == "__main__":
         for batt in range(lim):
             batt_modules.append(BeVolt.modules[ic * 8 + batt])
 
-        minions.append(LTC6811(batt_modules))
+        minions.append(ltc6811.LTC6811(batt_modules))
+    print()
 
     # Test PEC
-    LTC6811.pec_init()
-    print(hex(LTC6811.calc_pec([0x00, 0x01])))
+    print("PEC Test")
+    ltc6811.pec_init()
+    print(hex(ltc6811.calc_pec([0x00, 0x01])))
+    print()
 
     # Test LTC6811 static methods
+    print("LTC6811 Static Methods Test")
     minions[0].tx_data = [0, 1, 2, 3, 4, 5]
     minions[1].tx_data = [6, 7, 8, 9, 10, 11]
     minions[2].tx_data = [12, 13, 14, 15, 16, 17]
     minions[3].tx_data = [18, 19, 20, 21, 22, 23]
-    protocol = LTC6811.format_full_protocol(minions)
-    LTC6811.print_protocol(protocol)
+    protocol = ltc6811.format_full_protocol(minions)
+    ltc6811.print_protocol(protocol)
 
     # Add command at the front + cmd_pec
-    pec = LTC6811.calc_pec([0x00, 0x01])
+    pec = ltc6811.calc_pec([0x00, 0x01])
     protocol.insert(0, 0x00)
     protocol.insert(1, 0x01)
     protocol.insert(2, (pec >> 8) & 0x00FF)    # PEC0
     protocol.insert(3, pec & 0x00FF)           # PEC1
 
-    LTC6811.parse_full_protocol(minions, protocol)
+    ltc6811.parse_full_protocol(minions, protocol)
 
     for minion in minions:
         print(minion.rx_data)
